添加ping,pong

+++ common/src/main/java/com/example/code/TypeData.java
@@ -9,21 +9,21 @@ public interface TypeData {
     /**
      * 用户注册信息状态
      */
-    byte ACK = 0;
+    int ACK = 0;
 
     /**
      * ping消息状态
      */
-    byte PING = 1;
+    int PING = 1;
 
     /**
      * pong消息状态
      */
-    byte PONG = 2;
+    int PONG = 2;
     /**
      * 发送消息状态
      */
-    byte MESSAGE = 3;
+    int MESSAGE = 3;
 
 
 }
